requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.1
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
wheel==0.45.1
yarl==1.20.1
python-dotenv==1.1.1
quart==0.20.0
openai==1.95.1
certifi==2024.2.2